ya hice solicitudes
Acomode el backend de formulario
